+++ third-party/flycheck/CHANGES.rst
@@ -23,6 +23,9 @@
   - Rust with ``rust-clippy`` [GH-1385]
   - VHDL with ``ghdl`` [GH-1160]
   - mypy with ``python-mypy`` [GH-1354]
+  - Nix with ``nix-linter`` [GH-1530]
+  - Opam with ``opam lint`` [GH-1532]
+  - Text prose with ``textlint`` [GH-1534]
 
 - New features: